pyproject.toml
[tool.poetry]
name = "llm-app"
version = "0.1.0"
description = ""
authors = ["Aasim Khan <khan.aasim.88@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
langchain = "^0.3.3"
langchain-community = "^0.3.2"
langchain-core = "^0.3.10"
langchain-openai = "^0.2.2"
python-dotenv = "^1.0.1"
fastapi = {extras = ["standard"], version = "^0.115.2"}
uvicorn = "^0.32.0"
python-multipart = "^0.0.12"

[tool.poetry.group.dev.dependencies]
ruff = "^0.6.9"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
